requirements.txt
Flask
python-docx
Pillow
moviepy
gtts
